autooptimize: consolidate duplicate helper functions into text_utils.py (-80 LOC)
