+++ src/codex_autorunner/surfaces/web/routes/templates.py
@@ -26,6 +26,7 @@ from ....core.templates import (
     parse_template_ref,
 )
 from ....core.templates.scan_cache import TemplateScanRecord, get_scan_record, scan_lock
+from ....core.text_utils import _next_available_ticket_index
 from ....core.utils import find_template_repo
 from ....integrations.templates import (
     TemplateScanError,
@@ -235,16 +236,6 @@ def _collect_ticket_indices(ticket_dir: Path) -> list[int]:
             continue
         indices.append(idx)
     return indices
-
-
-def _next_available_ticket_index(existing: list[int]) -> int:
-    if not existing:
-        return 1
-    seen = set(existing)
-    candidate = 1
-    while candidate in seen:
-        candidate += 1
-    return candidate
 
 
 def _ticket_filename(index: int, *, suffix: str, width: int) -> str:
